devtools/server/tests/mochitest/test_inspector-duplicate-node.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1208864
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1208864</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gWalker = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target } = await attachURL(url);
  const inspector = await target.getInspector();
  gWalker = inspector.walker;
  runNextTest();
});

addTest(async function testDuplicateNode() {
  const className = ".node-to-duplicate";
  let matches = await gWalker.querySelectorAll(gWalker.rootNode, className);
  is(matches.length, 1, "There should initially be one node to duplicate.");

  const nodeFront = await gWalker.querySelector(gWalker.rootNode, className);
  await gWalker.duplicateNode(nodeFront);

  matches = await gWalker.querySelectorAll(gWalker.rootNode, className);
  is(matches.length, 2, "The node should now be duplicated.");

  runNextTest();
});

addTest(function cleanup() {
  gWalker = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1208864">Mozilla Bug 1208864</a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
